refactored byte swapping into utility class
